style: fix all pylint warnings and move to utils
